+++ doc/rst/MochiKit/VersionHistory.rst
@@ -1,7 +1,8 @@
 2005-XX-XX      v1.1
 
 - Fixed isoTimestamp to handle timestamps with time zone correctly
-- Added new MochiKit.DOM createDOMFunc aliases: SELECT, OPTION
+- Added new MochiKit.DOM createDOMFunc aliases: SELECT, OPTION, OPTGROUP, 
+  LEGEND, FIELDSET
 - New MochiKit.DOM formContents and enhancement to queryString to support it
 - Updated view_source example to use dp.SyntaxHighlighter 1.3.0
 - MochiKit.LoggingPane now uses named windows based on the URL so that